chore: split out code
